+++ typo3/sysext/fluid/ext_localconf.php
@@ -6,6 +6,7 @@ if (!defined('TYPO3_MODE')) {
 if (!is_array($GLOBALS['TYPO3_CONF_VARS']['SYS']['caching']['cacheConfigurations']['fluid_template'])) {
 	$GLOBALS['TYPO3_CONF_VARS']['SYS']['caching']['cacheConfigurations']['fluid_template'] = array(
 		'backend' => 'TYPO3\\CMS\\Core\\Cache\\Backend\\FileBackend',
-		'frontend' => 'TYPO3\\CMS\\Core\\Cache\\Frontend\\PhpFrontend'
+		'frontend' => 'TYPO3\\CMS\\Core\\Cache\\Frontend\\PhpFrontend',
+		'groups' => array('system')
 	);
 }